Dawood and Wadha task11

+++ restaurants/views.py
@@ -3,6 +3,7 @@ from .models import Restaurant, Item
 from .forms import RestaurantForm, ItemForm, SignupForm, SigninForm
 from django.contrib.auth import login, authenticate, logout
 
+
 def signup(request):
     form = SignupForm()
     if request.method == 'POST':
@@ -59,9 +60,12 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+    if request.user.is_anonymous:
+        return redirect('signin')
+
     form = RestaurantForm()
     if request.method == "POST":
-        form = RestaurantForm(request.POST, request.FILES)
+        form = RestaurantForm(request.POST, request.FILES or None)
         if form.is_valid():
             restaurant = form.save(commit=False)
             restaurant.owner = request.user
@@ -75,6 +79,10 @@ def restaurant_create(request):
 def item_create(request, restaurant_id):
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
+
+    if not (request.user.is_staff or request.user == restaurant.owner):
+        return redirect('no-access')
+
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -90,9 +98,11 @@ def item_create(request, restaurant_id):
 
 def restaurant_update(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+    if not (request.user.is_staff or request.user == restaurant_obj.owner):
+        return redirect('no-access')
     form = RestaurantForm(instance=restaurant_obj)
     if request.method == "POST":
-        form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
+        form = RestaurantForm(request.POST, request.FILES or None, instance=restaurant_obj)
         if form.is_valid():
             form.save()
             return redirect('restaurant-list')
@@ -103,6 +113,12 @@ def restaurant_update(request, restaurant_id):
     return render(request, 'update.html', context)
 
 def restaurant_delete(request, restaurant_id):
+    if not request.user.is_staff:
+        return redirect('no-access')
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     restaurant_obj.delete()
     return redirect('restaurant-list')
+
+
+def no_access(request):
+    return render(request,'no_access.html')
